Port unit tests to ShellSpec
The primary reason for this is that ShellSpec supports testing side effects of
shell functions, e.g. leaked variables or functions which store results in a
shell variable.
Doing this with bats would be non-trivial.

Futher ShellSpec does not rely on GNU Bash, executes tests faster, has more
readable tests.